rename commands
